Clean up models and add algorithms param to experiments endpoint
