+++ nectar/models/__init__.py
@@ -1,0 +1,7 @@
+from .experiment import Experiment
+from .gym_environment import GymEnvironment
+
+__all__ = [
+    "Experiment",
+    "GymEnvironment"
+]
