+++ nectar/models/gym_environment.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class GymEnvironment(BaseModel):
+    environmentName: str
+    environmentPath: str
